# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MedGama Telehealth API",
...
        "contact": {
            "name": "MedGama Dev Team",
            "email": "dev@medgama.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001/api",
            "description": "Local Development"
        },
        {
            "url": "https://api.medgama.com/api",
            "description": "Production"
        }
    ],
    "paths": {
...
        "/medstream/posts": {
            "get": {
                "tags": [
                    "MedStream"
                ],
                "summary": "List posts (public, paginated)",
                "operationId": "976dceb17cadf5c93b142e3a32b5066b",
                "parameters": [
                    {
                        "name": "author_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "clinic_id",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "post_type",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "article",
                                "case",
                                "question",
                                "media"
                            ]
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated posts (MedStreamPostResource)"
                    }
                }
            },
            "post": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Create post (doctor/clinicOwner/admin). Videos processed asynchronously.",
                "operationId": "69468c76b260b279beddbdd753e38b61",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "post_type",
                                    "content"
                                ],
                                "properties": {
                                    "post_type": {
                                        "type": "string",
                                        "enum": [
                                            "article",
                                            "case",
                                            "question",
                                            "media"
                                        ]
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "clinic_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "photos[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "videos[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created (media_processing may be true for videos)"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/posts/{post}": {
            "get": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Show post detail with comments and engagement",
                "operationId": "c9c0a9a42ff5af5dbb62bcaf5620ecb3",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post detail (MedStreamPostResource)"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Update post (author or admin)",
                "operationId": "03c52342b8ebbdd1bd940e0228a59a70",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post updated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Delete post (author or admin)",
                "operationId": "297459f8ba7bdefe11e61c306bc39579",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/posts/{post}/comments": {
            "get": {
                "tags": [
                    "MedStream"
                ],
                "summary": "List comments for a post",
                "operationId": "0c9bc18168f13369b604681774946181",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated comments (MedStreamCommentResource)"
                    }
                }
            },
            "post": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Add comment (supports threaded replies via parent_id)",
                "operationId": "3058edc560c43023e4426b698752da9f",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "description": "For threaded replies",
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Comment created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/posts/{post}/like": {
            "post": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Toggle like on post",
                "operationId": "3d5d86a7c214341c5495a3c41bed9edb",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Like toggled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/bookmarks": {
            "get": {
                "tags": [
                    "MedStream"
                ],
                "summary": "List bookmarked posts",
                "operationId": "432252f07c4f407c860e1b1d5b31e5fa",
                "responses": {
                    "200": {
                        "description": "Paginated bookmarks"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Toggle bookmark",
                "operationId": "0855b91a7125f5eea18257265ecb3704",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bookmarked_type",
                                    "target_id"
                                ],
                                "properties": {
                                    "bookmarked_type": {
                                        "type": "string"
                                    },
                                    "target_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Bookmark toggled"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/posts/{post}/report": {
            "post": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Report a post",
                "operationId": "d80cd9e662e4905c6989e9d632cf92cf",
                "parameters": [
                    {
                        "name": "post",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/reports": {
            "get": {
                "tags": [
                    "MedStream"
                ],
                "summary": "List reports (admin only)",
                "operationId": "1b66d2bc33ad311de30a6087eba4da12",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated reports"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/medstream/reports/{id}": {
            "put": {
                "tags": [
                    "MedStream"
                ],
                "summary": "Update report status (admin only)",
                "operationId": "cca3e6481dfc0b428f8fa504efc6d4f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Report updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}